Event_Registration/requirements.txt
django>=4.2
djangorestframework
PyMySQL
python-dotenv
